Split ReportVacation into Edit and Create
