+++ Presentation.Web/App/Core/Interfaces/KendoGrid.d.ts
@@ -1,5 +1,5 @@
 ﻿declare namespace  app.core.interfaces {
     export interface KendoGrid extends kendo.ui.Grid {
-        dataSource: any; // This is the type kendo uses in GridOptions
+        dataSource: any;
     }
 }
