+++ Presentation.Web/App/Vacation/ReportVacation/CreateReportVacationController.ts
@@ -1,0 +1,88 @@
+﻿module app.vacation {
+    "use strict";
+
+    import VacationBalanceResource = vacation.resources.IVacationBalanceResource;
+    import Person = core.models.Person;
+    import NotificationService = core.interfaces.NotificationService;
+
+    class CreateReportVacationController extends BaseReportVacationController {
+
+        static $inject = [
+            "$scope",
+            "Person",
+            "$rootScope",
+            "VacationReport",
+            "NotificationService",
+            "VacationBalanceResource",
+            "moment"
+        ];
+
+        constructor(
+            protected $scope,
+            protected Person: Person,
+            protected $rootScope,
+            protected VacationReport, // TODO Make $resource class
+            protected NotificationService: NotificationService,
+            protected VacationBalanceResource: VacationBalanceResource,
+            protected moment: moment.MomentStatic) {
+
+            super($scope, Person, $rootScope, VacationReport, NotificationService, VacationBalanceResource, moment);
+
+            this.initializeReport();
+        }
+
+        protected initializeReport() {
+            this.startDate = new Date();
+            this.endDate = new Date();
+            this.maxEndDate = new Date();
+            this.comment = undefined;
+            this.vacationStartsOnFullDay = true;
+            this.vacationEndsOnFullDay = true;
+            this.startTime = new Date(2000, 0, 1, 0, 0, 0, 0); // Only time is relevant, date is ignored by kendo
+            this.endTime = new Date(2000, 0, 1, 0, 0, 0, 0);
+            this.vacationType = undefined;
+        }
+
+        saveReport() {
+            const report = new this.VacationReport();
+
+            report.StartTimestamp = Math.floor((new Date(this.startDate.getFullYear(), this.startDate.getMonth(), this.startDate.getDate()).getTime()) / 1000);
+            report.EndTimestamp = Math.floor((new Date(this.endDate.getFullYear(), this.endDate.getMonth(), this.endDate.getDate()).getTime()) / 1000);
+            report.EmploymentId = this.position;
+            report.Comment = this.comment;
+            report.PersonId = this.currentUser.Id;
+            report.Status = "Pending";
+            report.CreatedDateTimestamp = Math.floor(Date.now() / 1000);
+            report.VacationType = this.vacationType;
+
+            if (!this.vacationStartsOnFullDay) {
+                report.StartTime = `P0DT${this.startTime.getHours()}H${this.startTime.getMinutes()}M0S`;
+            } else {
+                report.StartTime = null;
+            }
+
+            if (!this.vacationEndsOnFullDay) {
+                report.EndTime = `P0DT${this.endTime.getHours()}H${this.endTime.getMinutes()}M0S`;
+            } else {
+                report.EndTime = null;
+            }
+
+            report.$save(() => {
+                this.NotificationService
+                    .AutoFadeNotification("success", "", "Din indberetning er sendt til godkendelse.");
+
+                this.clearReport();
+                this.saveButtenDisabled = false;
+            },
+            () => {
+                this.saveButtenDisabled = false;
+                this.NotificationService
+                    .AutoFadeNotification("danger",
+                        "",
+                        "Der opstod en fejl under oprettelsen af din ferieindberetning.");
+            });
+        }
+    }
+
+    angular.module("app.vacation").controller("CreateReportVacationController", CreateReportVacationController);
+}
